+++ pkg/services/ngalert/api/lotex.go
@@ -2,7 +2,6 @@ package api
 
 import (
 	"bytes"
-	"encoding/json"
 	"fmt"
 	"io"
 	"io/ioutil"
@@ -10,96 +9,25 @@ import (
 	"net/url"
 
 	apimodels "github.com/grafana/alerting-api/pkg/api"
-	"gopkg.in/macaron.v1"
 	"gopkg.in/yaml.v3"
 
 	"github.com/grafana/grafana/pkg/api/response"
 	"github.com/grafana/grafana/pkg/infra/log"
 	"github.com/grafana/grafana/pkg/models"
-	"github.com/grafana/grafana/pkg/services/datasourceproxy"
 )
 
 const legacyRulerPrefix = "/api/prom/rules"
 
 type LotexRuler struct {
-	DataProxy *datasourceproxy.DatasourceProxyService
-	log       log.Logger
+	log log.Logger
+	*AlertingProxy
 }
 
-// macaron unsafely asserts the http.ResponseWriter is an http.CloseNotifier, which will panic.
-// Here we impl it, which will ensure this no longer happens, but neither will we take
-// advantage cancelling upstream requests when the downstream has closed.
-// NB: http.CloseNotifier is a deprecated ifc from before the context pkg.
-type safeMacaronWrapper struct {
-	http.ResponseWriter
-}
-
-func (w *safeMacaronWrapper) CloseNotify() <-chan bool {
-	return make(chan bool)
-}
-
-// replacedResponseWriter overwrites the underlying responsewriter used by a *models.ReqContext.
-// It's ugly because it needs to replace a value behind a few nested pointers.
-func replacedResponseWriter(ctx *models.ReqContext) (*models.ReqContext, *response.NormalResponse) {
-	resp := response.CreateNormalResponse(make(http.Header), nil, 0)
-	cpy := *ctx
-	cpyMCtx := *cpy.Context
-	cpyMCtx.Resp = macaron.NewResponseWriter(ctx.Req.Method, &safeMacaronWrapper{resp})
-	cpy.Context = &cpyMCtx
-	return &cpy, resp
-}
-
-// withReq proxies a different request
-func (r *LotexRuler) withReq(
-	ctx *models.ReqContext,
-	req *http.Request,
-	extractor func([]byte) (interface{}, error),
-) response.Response {
-	newCtx, resp := replacedResponseWriter(ctx)
-	newCtx.Req.Request = req
-	r.DataProxy.ProxyDatasourceRequestWithID(newCtx, ctx.ParamsInt64("Recipient"))
-
-	status := resp.Status()
-	if status >= 400 {
-		return response.Error(status, string(resp.Body()), nil)
+func NewLotexRuler(proxy *AlertingProxy, log log.Logger) *LotexRuler {
+	return &LotexRuler{
+		log:           log,
+		AlertingProxy: proxy,
 	}
-
-	t, err := extractor(resp.Body())
-	if err != nil {
-		return response.Error(500, err.Error(), nil)
-	}
-
-	b, err := json.Marshal(t)
-	if err != nil {
-		return response.Error(500, err.Error(), nil)
-	}
-
-	return response.JSON(status, b)
-}
-
-func yamlExtractor(v interface{}) func([]byte) (interface{}, error) {
-	return func(b []byte) (interface{}, error) {
-		decoder := yaml.NewDecoder(bytes.NewReader(b))
-		decoder.KnownFields(true)
-
-		err := decoder.Decode(v)
-
-		return v, err
-	}
-}
-
-func jsonExtractor(v interface{}) func([]byte) (interface{}, error) {
-	if v == nil {
-		// json unmarshal expects a pointer
-		v = &map[string]interface{}{}
-	}
-	return func(b []byte) (interface{}, error) {
-		return v, json.Unmarshal(b, v)
-	}
-}
-
-func messageExtractor(b []byte) (interface{}, error) {
-	return map[string]string{"message": string(b)}, nil
 }
 
 func (r *LotexRuler) RouteDeleteNamespaceRulesConfig(ctx *models.ReqContext) response.Response {
